requirements.txt
streamlit
langchain
langchain-community
openai
PyPDF2
python-docx
Pillow
faiss-cpu
requests
tiktoken